cbloader delegates split into separate files into _delegates folder
