+++ avalon/tools/_delegates/delegate_pretty_time.py
@@ -1,0 +1,13 @@
+from . import QtWidgets
+from .lib import pretty_timestamp
+
+
+class PrettyTimeDelegate(QtWidgets.QStyledItemDelegate):
+    """A delegate that displays a timestamp as a pretty date.
+
+    This displays dates like `pretty_date`.
+
+    """
+
+    def displayText(self, value, locale):
+        return pretty_timestamp(value)
